Separate SliderImageCard into SliderImage and SliderImageCard components
