+++ src/components/SliderImageCard.tsx
@@ -1,10 +1,10 @@
-import { MouseEvent, useState, ReactNode } from "react";
+import { ReactNode } from "react";
 import ElementCard, { Props as ElementCardProps } from "components/ElementCard";
-import { SliderImage } from "helpers/customTypes";
-import DefaultPic from "assets/default-pic.png";
+import { SliderImage as SliderImageType } from "helpers/customTypes";
+import SliderImage from "components/SliderImage";
 
 type Props = {
-  images: SliderImage[];
+  images: (SliderImageType & { src?: string })[];
   children?: ReactNode;
   imageClassName?: string;
 } & Omit<ElementCardProps, "children">;
@@ -15,70 +15,14 @@ export default function SliderImageCard({
   imageClassName = "",
   ...otherProps
 }: Props) {
-  const [imageIndex, setImageIndex] = useState(0);
-  const handleNextImage = (event: MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault();
-    setImageIndex((currentIndex) =>
-      currentIndex === images.length - 1 ? 0 : currentIndex + 1
-    );
-  };
-
-  const handlePrevImage = (event: MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault();
-    setImageIndex((currentIndex) =>
-      currentIndex === 0 ? images.length - 1 : currentIndex - 1
-    );
-  };
-
   return (
     <ElementCard type={type} {...otherProps}>
-      {images.length > 1 && (
-        <button
-          onClick={handlePrevImage}
-          className={
-            "absolute left-2 " +
-            "text-slate-500 " +
-            "h-full w-5 " +
-            "hover:text-slate-900 hover:text-lg"
-          }
-          data-testid="prev-image_button"
-          aria-label="imagen anterior"
-        >
-          <i className="fa-solid fa-angle-left" />
-        </button>
-      )}
-      {children && children}
-      <div className="w-full flex items-center justify-center">
-        <img
-          data-testid={`slider-image_${images[imageIndex]?.id || "default"}`}
-          className={`${imageClassName} ` + "max-h-64"}
-          alt={
-            images[imageIndex]?.id
-              ? `${type} image ${images[imageIndex].id}`
-              : "default image"
-          }
-          src={
-            images[imageIndex]?.name
-              ? `${import.meta.env.VITE_BASE_URL}/${images[imageIndex]?.name}`
-              : DefaultPic
-          }
-        />
-      </div>
-      {images.length > 1 && (
-        <button
-          onClick={handleNextImage}
-          className={
-            "absolute right-2 " +
-            "text-slate-500 " +
-            "h-full w-5 " +
-            "hover:text-slate-900 hover:text-lg "
-          }
-          data-testid="next-image_button"
-          aria-label="imagen siguiente"
-        >
-          <i className="fa-solid fa-angle-right" />
-        </button>
-      )}
+      <SliderImage
+        images={images}
+        children={children}
+        imageClassName={imageClassName}
+        type={type}
+      />
     </ElementCard>
   );
 }
